requirements.txt
pydantic-ai>=0.0.13
google-generativeai>=0.8.0
asyncpg>=0.29.0
sqlmodel>=0.0.22
python-dotenv>=1.0.0
pydantic>=2.0.0
